+++ Providers/AspNet/Revo.AspNet/Security/SecurityModule.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin.Security;
 using Ninject.Modules;
-using Revo.AspNet.Security.Identity;
+using Revo.AspNet.Web;
 using Revo.Core.Core;
 using Revo.Core.Security;
 
@@ -15,14 +15,6 @@ namespace Revo.AspNet.Security
     {
         public override void Load()
         {
-            Bind<SignInManager<IIdentityUser, Guid>>()
-                .To<SignInManager<IIdentityUser, Guid>>()
-                .InTaskScope();
-
-            Bind<UserManager<IIdentityUser, Guid>>()
-                .To<UserManager<IIdentityUser, Guid>>()
-                .InTaskScope();
-
             Bind<IAuthenticationManager>()
                 .ToMethod(ctx =>
                     HttpContext.Current?.TryGetOwinContext()?.Authentication);