+++ Providers/AspNetCore/Revo.AspNetCore/Security/Identity/ISignInManager.cs
@@ -1,9 +1,0 @@
-﻿using System.Security.Claims;
-
-namespace Revo.AspNetCore.Security.Identity
-{
-    public interface ISignInManager
-    {
-        bool IsSignedIn(ClaimsPrincipal principal);
-    }
-}