+++ Revo.Core/Security/IUserPermissionAuthorizer.cs
@@ -1,4 +1,6 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace Revo.Core.Security
 {
@@ -11,12 +13,24 @@ namespace Revo.Core.Security
         /// Checks authorization for an action of specified user, requiring specified permission
         /// and (optionally) resource or context.
         /// </summary>
-        Task<bool> CheckAuthorizationAsync(IUser user, string permissionId, string resourceId = null, string contextId = null);
+        Task<bool> CheckAuthorizationAsync(IUser user, Guid permissionId, string resourceId = null, string contextId = null);
+
+        /// <summary>
+        /// Checks authorization for an action of specified user, requiring specified permission
+        /// and (optionally) resource or context.
+        /// </summary>
+        Task<bool> CheckAuthorizationAsync(IUser user, IEnumerable<Permission> permissions);
 
         /// <summary>
         /// Checks authorization for an action of the current user (as provided by IUserContext),
         /// requiring specified permission and (optionally) resource or context.
         /// </summary>
-        Task<bool> CheckCurrentUserAuthorizationAsync(string permissionId, string resourceId = null, string contextId = null);
+        Task<bool> CheckCurrentUserAuthorizationAsync(Guid permissionId, string resourceId = null, string contextId = null);
+
+        /// <summary>
+        /// Checks authorization for an action of the current user (as provided by IUserContext),
+        /// requiring specified permission and (optionally) resource or context.
+        /// </summary>
+        Task<bool> CheckCurrentUserAuthorizationAsync(IEnumerable<Permission> permissions);
     }
 }