+++ Revo.Core/Security/NullCoreSecurityModule.cs
@@ -1,22 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Ninject.Modules;
+﻿using Ninject.Modules;
+using Revo.Core.Core;
+using Revo.Core.Security.ClaimBased;
 
 namespace Revo.Core.Security
 {
+    [AutoLoadModule(false)]
     public class NullCoreSecurityModule : NinjectModule
     {
         public override void Load()
         {
+            Bind<IClaimsPrincipalUserResolver>()
+                .To<NullClaimsPrincipalUserResolver>()
+                .InSingletonScope();
+
             Bind<IUserPermissionResolver>()
                 .To<NullUserPermissionResolver>()
                 .InSingletonScope();
-
-            Bind<IRolePermissionResolver>()
-                .To<NullRolePermissionResolver>()
-                .InSingletonScope();
         }
-
     }
 }