+++ Revo.Core/Security/UserPermissionAuthorizer.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace Revo.Core.Security
@@ -21,24 +22,38 @@ namespace Revo.Core.Security
             this.permissionTypeRegistry = permissionTypeRegistry;
         }
 
-        public async Task<bool> CheckAuthorizationAsync(IUser user, string permissionId, string resourceId = null, string contextId = null)
+        public async Task<bool> CheckAuthorizationAsync(IUser user, Guid permissionId, string resourceId = null, string contextId = null)
         {
             var userPermissions = await userPermissionResolver.GetUserPermissionsAsync(user);
             return permissionAuthorizationMatcher.CheckAuthorization(userPermissions,
                 new[] { GetPermission(permissionId, resourceId, contextId) });
         }
 
-        public async Task<bool> CheckCurrentUserAuthorizationAsync(string permissionId, string resourceId = null, string contextId = null)
+        public async Task<bool> CheckAuthorizationAsync(IUser user, IEnumerable<Permission> permissions)
+        {
+            var userPermissions = await userPermissionResolver.GetUserPermissionsAsync(user);
+            return permissionAuthorizationMatcher.CheckAuthorization(userPermissions,
+                permissions);
+        }
+
+        public async Task<bool> CheckCurrentUserAuthorizationAsync(Guid permissionId, string resourceId = null, string contextId = null)
         {
             var userPermissions = await userContext.GetPermissionsAsync();
             return permissionAuthorizationMatcher.CheckAuthorization(userPermissions,
                 new[] { GetPermission(permissionId, resourceId, contextId) });
         }
 
-        private Permission GetPermission(string permissionId, string resourceId, string contextId)
+        public async Task<bool> CheckCurrentUserAuthorizationAsync(IEnumerable<Permission> permissions)
+        {
+            var userPermissions = await userContext.GetPermissionsAsync();
+            return permissionAuthorizationMatcher.CheckAuthorization(userPermissions,
+                permissions);
+        }
+
+        private Permission GetPermission(Guid permissionId, string resourceId, string contextId)
         {
             return new Permission(
-                permissionTypeRegistry.GetPermissionTypeById(Guid.Parse(permissionId)),
+                permissionTypeRegistry.GetPermissionTypeById(permissionId),
                 resourceId,
                 contextId);
         }